添加ssr代码

+++ engine/source/runtime/function/render/rhi/d3d12/d3d12_linkedDevice.cpp
@@ -31,7 +31,7 @@ static int CVar_GlobalDynamicGPUSamplerHeapSize = 1024;
 
 namespace RHI
 {
-    UINT D3D12LinkedDevice::m_FrameIndex = 0;
+    long D3D12LinkedDevice::m_FrameIndex = 0;
 
     // clang-format off
     D3D12LinkedDevice::D3D12LinkedDevice(D3D12Device* Parent, D3D12NodeMask NodeMask)
