+++ engine/source/runtime/function/render/rhi/d3d12/d3d12_linkedDevice.h
@@ -71,7 +71,7 @@ namespace RHI
         void Release(D3D12SyncHandle syncHandle);
         //-----------------------------------------------------------
 
-        static UINT m_FrameIndex;
+        static long m_FrameIndex;
     private:
         static constexpr UINT MaxSharedBufferCount = 3;
